fix: conv type

+++ MaxText/llama_mistral_mixtral_orbax_to_hf.py
@@ -137,7 +137,7 @@ def convert_state_to_hf(training_state, hf_model, model_size):
   print('param and shape: ', jax.tree.map(lambda x: x.shape, training_state.params))
   # Port the embedding weights
   hf_model_params["model.embed_tokens.weight"] = torch.tensor(
-      np.asarray(training_state.params["params"]["token_embedder"]["embedding"]), dtype=hf_model.get_parameter("model.embed_tokens.weight").dtype
+      np.asarray(training_state.params["params"]["token_embedder"]["embedding"].astype('float32')), dtype=hf_model.get_parameter("model.embed_tokens.weight").dtype
   )
 
   for layer_int in tqdm(range(base_num_decoder_layers), desc="Porting parameters layerwise"):
@@ -156,7 +156,7 @@ def convert_state_to_hf(training_state, hf_model, model_size):
                 model_size,
             )
             .reshape(hidden_size, base_num_query_heads * head_dim)
-            .T
+            .T.astype('float32')
         ),
         dtype=hf_model.get_parameter(f"model.layers.{layer_int}.self_attn.q_proj.weight").dtype,
     )
@@ -168,7 +168,7 @@ def convert_state_to_hf(training_state, hf_model, model_size):
                 model_size,
             )
             .reshape(hidden_size, base_num_kv_heads * head_dim)
-            .T
+            .T.astype('float32')
         ),
         dtype=hf_model.get_parameter(f"model.layers.{layer_int}.self_attn.k_proj.weight").dtype,
     )
@@ -176,7 +176,7 @@ def convert_state_to_hf(training_state, hf_model, model_size):
         np.asarray(
             training_state.params["params"]["decoder"]["layers"]["self_attention"]["value"]["kernel"][:, layer_int, :, :]
             .reshape(hidden_size, base_num_kv_heads * head_dim)
-            .T
+            .T.astype('float32')
         ),
         dtype=hf_model.get_parameter(f"model.layers.{layer_int}.self_attn.v_proj.weight").dtype,
     )
@@ -184,7 +184,7 @@ def convert_state_to_hf(training_state, hf_model, model_size):
         np.asarray(
             training_state.params["params"]["decoder"]["layers"]["self_attention"]["out"]["kernel"][:, layer_int, :, :]
             .reshape(base_num_query_heads * head_dim, hidden_size)
-            .T
+            .T.astype('float32')
         ),
         dtype=hf_model.get_parameter(f"model.layers.{layer_int}.self_attn.o_proj.weight").dtype,
     )
@@ -192,35 +192,35 @@ def convert_state_to_hf(training_state, hf_model, model_size):
     # MLP Layers
     if num_experts is None:
       hf_model_params[f"model.layers.{layer_int}.mlp.gate_proj.weight"] = torch.tensor(
-          np.asarray(training_state.params["params"]["decoder"]["layers"]["mlp"]["wi_0"]["kernel"][:, layer_int, :].T),
+          np.asarray(training_state.params["params"]["decoder"]["layers"]["mlp"]["wi_0"]["kernel"][:, layer_int, :].T.astype('float32')),
           dtype=hf_model.get_parameter(f"model.layers.{layer_int}.mlp.gate_proj.weight").dtype,
       )
       hf_model_params[f"model.layers.{layer_int}.mlp.up_proj.weight"] = torch.tensor(
-          np.asarray(training_state.params["params"]["decoder"]["layers"]["mlp"]["wi_1"]["kernel"][:, layer_int, :].T),
+          np.asarray(training_state.params["params"]["decoder"]["layers"]["mlp"]["wi_1"]["kernel"][:, layer_int, :].T.astype('float32')),
           dtype=hf_model.get_parameter(f"model.layers.{layer_int}.mlp.up_proj.weight").dtype,
       )
       hf_model_params[f"model.layers.{layer_int}.mlp.down_proj.weight"] = torch.tensor(
-          np.asarray(training_state.params["params"]["decoder"]["layers"]["mlp"]["wo"]["kernel"][:, layer_int, :].T),
+          np.asarray(training_state.params["params"]["decoder"]["layers"]["mlp"]["wo"]["kernel"][:, layer_int, :].T.astype('float32')),
           dtype=hf_model.get_parameter(f"model.layers.{layer_int}.mlp.down_proj.weight").dtype,
       )
     else:
       hf_model_params[f"model.layers.{layer_int}.block_sparse_moe.gate.weight"] = torch.tensor(
           np.asarray(
-              training_state.params["params"]["decoder"]["layers"]["MoeBlock_0"]["gate"]["kernel"][:, layer_int, :].T
+              training_state.params["params"]["decoder"]["layers"]["MoeBlock_0"]["gate"]["kernel"][:, layer_int, :].T.astype('float32')
           ),
           dtype=hf_model.get_parameter(f"model.layers.{layer_int}.block_sparse_moe.gate.weight").dtype,
       )
       for k in range(num_experts):
         hf_model_params[f"model.layers.{layer_int}.block_sparse_moe.experts.{k}.w1.weight"] = torch.tensor(
-            np.asarray(training_state.params["params"]["decoder"]["layers"]["MoeBlock_0"]["wi_0"][k, layer_int, :, :].T),
+            np.asarray(training_state.params["params"]["decoder"]["layers"]["MoeBlock_0"]["wi_0"][k, layer_int, :, :].T.astype('float32')),
             dtype=hf_model.get_parameter(f"model.layers.{layer_int}.block_sparse_moe.experts.{k}.w1.weight").dtype,
         )
         hf_model_params[f"model.layers.{layer_int}.block_sparse_moe.experts.{k}.w2.weight"] = torch.tensor(
-            np.asarray(training_state.params["params"]["decoder"]["layers"]["MoeBlock_0"]["wo"][k, layer_int, :, :].T),
+            np.asarray(training_state.params["params"]["decoder"]["layers"]["MoeBlock_0"]["wo"][k, layer_int, :, :].T.astype('float32')),
             dtype=hf_model.get_parameter(f"model.layers.{layer_int}.block_sparse_moe.experts.{k}.w2.weight").dtype,
         )
         hf_model_params[f"model.layers.{layer_int}.block_sparse_moe.experts.{k}.w3.weight"] = torch.tensor(
-            np.asarray(training_state.params["params"]["decoder"]["layers"]["MoeBlock_0"]["wi_1"][k, layer_int, :, :].T),
+            np.asarray(training_state.params["params"]["decoder"]["layers"]["MoeBlock_0"]["wi_1"][k, layer_int, :, :].T.astype('float32')),
             dtype=hf_model.get_parameter(f"model.layers.{layer_int}.block_sparse_moe.experts.{k}.w3.weight").dtype,
         )
 
@@ -229,7 +229,7 @@ def convert_state_to_hf(training_state, hf_model, model_size):
         np.asarray(
             training_state.params["params"]["decoder"]["layers"]["pre_self_attention_layer_norm"]["scale"][
                 :, layer_int
-            ].reshape(base_num_query_heads * head_dim)
+            ].reshape(base_num_query_heads * head_dim).astype('float32')
         ),
         dtype=hf_model.get_parameter(f"model.layers.{layer_int}.input_layernorm.weight").dtype,
     )
@@ -237,18 +237,18 @@ def convert_state_to_hf(training_state, hf_model, model_size):
         np.asarray(
             training_state.params["params"]["decoder"]["layers"]["post_self_attention_layer_norm"]["scale"][
                 :, layer_int
-            ].reshape(base_num_query_heads * head_dim)
+            ].reshape(base_num_query_heads * head_dim).astype('float32')
         ),
         dtype=hf_model.get_parameter(f"model.layers.{layer_int}.post_attention_layernorm.weight").dtype,
     )
 
   # LM head and layernorm
   hf_model_params["lm_head.weight"] = torch.tensor(
-      np.asarray(training_state.params["params"]["decoder"]["logits_dense"]["kernel"].T), dtype=hf_model.get_parameter("lm_head.weight").dtype
+      np.asarray(training_state.params["params"]["decoder"]["logits_dense"]["kernel"].T.astype('float32')), dtype=hf_model.get_parameter("lm_head.weight").dtype
   )
   hf_model_params["model.norm.weight"] = torch.tensor(
       np.asarray(
-          training_state.params["params"]["decoder"]["decoder_norm"]["scale"].reshape(base_num_query_heads * head_dim)
+          training_state.params["params"]["decoder"]["decoder_norm"]["scale"].reshape(base_num_query_heads * head_dim).astype('float32')
       ),
       dtype=hf_model.get_parameter("model.norm.weight").dtype,
   )
